Move /api/internal/rest to /api/rest so types can be consumed by external callers
